+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-11-21
+
+### Breaking changes
+
+- Don't pollute the command space:
+  Use a single command with subcommands and completions.
+  - `RustAnalyzer [start|stop]` 
+    (always available in Rust projects)
+  - `RustLsp [subcommand]` 
+    (available when the LSP client is running)
+    e.g. `RustLsp moveItem [up|down]`
+
 ## [1.0.1] - 2023-11-21
 
 ### Fixed